requirements.txt
alembic==1.12.1
asn1crypto==1.5.1
async-timeout==5.0.1
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==18.13.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.2.4
MarkupSafe==2.1.5
msgspec==0.19.0
packaging==24.0
pg8000==1.29.8
pluggy==1.2.0
psycopg==3.1.20
psycopg2-binary==2.9.10
pytest==7.4.4
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
redis==5.2.1
scramp==1.4.4
six==1.17.0
SQLAlchemy==2.0.36
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2025.1
Werkzeug==3.1.3
WTForms==3.0.1
zipp==3.15.0
